Correcion tabla de productos rechazados, finalizacion parcial de modulo de inventario, con ajustes pendientes en los submodulos de materias primas, productos rechazados y perfiles

+++ talcos-express/models/ProductosRechazados.js
@@ -15,7 +15,7 @@ const ProductosRechazados = sequelize.define(
       allowNull: false,
     },
     cantidad_producto_rechazado: {
-      type: DataTypes.BIGINT,
+      type: DataTypes.DECIMAL(5, 2),
       allowNull: false,
     },
     retencion_producto_rechazado: {
